opti: 按钮点按反馈优化&视频历史进度自动定位（仅限内建播放器

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/rcmd/Adver.kt
@@ -2,7 +2,7 @@ package cn.spacexc.wearbili.dataclass.video.rcmd
 
 data class Adver(
     val adver_desc: String,
-    val adver_id: Int,
+    val adver_id: Long,
     val adver_logo: String,
     val adver_name: String,
     val adver_page_url: String,
